+++ doajtest/unit/test_query.py
@@ -647,7 +647,11 @@ class TestQuery(DoajTestCase):
         assert res['hits']['total']["value"] == 1, res['hits']['total']["value"]
 
     def test_journal_query_ascii_folding(self):
-        self.journal = models.Journal(**JournalFixtureFactory.make_journal_with_data(title="I can’t really think in English"))
+        self.journal = JournalFixtureFactory.make_legacy_journal_object(in_doaj=True, overlay={
+            "bibjson": {
+                "title": "I can’t really think in English"
+            }
+        })
         self.journal.save(blocking=True)
         qsvc = QueryService()
 
@@ -709,11 +713,16 @@ class TestQuery(DoajTestCase):
         assert res['hits']['total']["value"] == 1, res['hits']['total']["value"]
 
     def test_journal_query_ascii_folding_data(self):
-        self.journal = models.Journal(**JournalFixtureFactory
-                                      .make_journal_with_data(title="Kadınlarının sağlık",
-                                                              publisher_name="Ankara Üniversitesi",
-                                                              country="Türkiye",
-                                                              alternative_title="Dirasat: Shariía and Law Sciences"))
+        self.journal = JournalFixtureFactory.make_legacy_journal_object(in_doaj=True, overlay={
+            "bibjson": {
+                "title": "Kadınlarının sağlık",
+                "publisher": {
+                    "name": "Ankara Üniversitesi",
+                    "country": "Türkiye"
+                },
+                "alternative_title": "Dirasat: Shariía and Law Sciences"
+            }
+        })
         self.journal.save(blocking=True)
         qsvc = QueryService()
 
@@ -748,10 +757,15 @@ class TestQuery(DoajTestCase):
 
     def test_application_query_ascii_folding_data(self):
         acc = models.Account(**AccountFixtureFactory.make_managing_editor_source())
-        application = models.Application(**ApplicationFixtureFactory
-                                      .make_application_with_data(title="Kadınlarının sağlık",
-                                                              publisher_name="Ankara Üniversitesi",
-                                                              country="Türkiye", ))
+        application = ApplicationFixtureFactory.make_legacy_application_object(overlay={
+            "bibjson": {
+                "title": "Kadınlarının sağlık",
+                "publisher": {
+                    "name": "Ankara Üniversitesi",
+                    "country": "Türkiye"
+                }
+            }
+        })
         application.save(blocking=True)
         qsvc = QueryService()