+++ portality/bll/services/events.py
@@ -27,6 +27,7 @@ from portality.events.consumers.journal_discontinuing_soon_notify import Journal
 from portality.events.consumers.application_editor_acceptreject_notify import ApplicationEditorAcceptRejectNotify
 from portality.events.consumers.update_request_maned_editor_group_assigned_notify import UpdateRequestManedEditorGroupAssignedNotify
 from portality.events.consumers.article_ris_generator import ArticleRISGenerator
+from portality.events.consumers.flag_assigned import FlagAssigned
 
 
 
@@ -55,6 +56,7 @@ class EventsService(object):
         ApplicationPublisherCreatedNotify,
         ApplicationPublisherQuickRejectNotify,
         ArticleRISGenerator,
+        FlagAssigned,
         BGJobFinishedNotify,
         JournalDiscontinuingSoonNotify,
         UpdateRequestManedEditorGroupAssignedNotify,