+++ portality/scripts/approaching_flag_deadline.py
@@ -1,0 +1,16 @@
+from portality.core import app
+from portality.tasks import approaching_flag_deadline
+from portality.background import BackgroundApi
+
+def main():
+    if app.config.get("SCRIPTS_READ_ONLY_MODE", False):
+        print("System is in READ-ONLY mode, script cannot run")
+        exit()
+
+    user = app.config.get("SYSTEM_USERNAME")
+    job = approaching_flag_deadline.ApproachingFlagDeadlineTask.prepare(user)
+    task = approaching_flag_deadline.ApproachingFlagDeadlineTask(job)
+    BackgroundApi.execute(task)
+
+if __name__ == "__main__":
+    main()