+++ portality/ui/templates.py
@@ -59,6 +59,7 @@ NOTIFICATIONS = "management/admin/notifications.html"
 ADMIN_UNLOCKED = "management/admin/unlocked.html"
 ADMIN_JOURNAL_CSV_SEARCH = "management/admin/journal_csv_search.html"
 ADMIN_PDD_SEARCH = "management/admin/pdd_search.html"
+ADMIN_SYSTEM_OBJECTS_SEARCH = "management/admin/system_objects/_search.html"
 
 # Application Form
 MANED_APPLICATION_FORM = "management/admin/maned_application.html"
@@ -80,7 +81,6 @@ AF_ENTRY_GROUP_HORIZONTAL = "_application-form/includes/_entry_group_horizontal.
 AF_FIELD = "_application-form/includes/_field.html"
 AF_GROUP = "_application-form/includes/_group.html"
 AF_LIST = "_application-form/includes/_list.html"
-FLAGS_LIST= "_application-form/includes/_flags_list.html"
 FLAG_ENTRY_GROUP = "_application-form/includes/_flag_entry.html"
 
 # Publisher area